+++ javascript/modules/client/src/internal/client-core.ts
@@ -5,6 +5,8 @@ import { Contract, ContractInterface } from "@ethersproject/contracts";
 import { IClientCore } from "./interfaces/client-core";
 import { Context } from "../context";
 import {
+  DaoDepositSteps,
+  DaoDepositStepValue,
   ICreateProposal,
   IDeposit,
   IGasFeeEstimation,
@@ -254,7 +256,9 @@ export abstract class ClientCore implements IClientCore {
 
   // DAO METHODS
 
-  protected async deposit(params: IDeposit): Promise<void> {
+  protected async *deposit(
+    params: IDeposit
+  ): AsyncGenerator<DaoDepositStepValue> {
     if (!this.connectedSigner) {
       throw new Error("A signer is needed for creating a DAO");
     }
@@ -265,6 +269,52 @@ export abstract class ClientCore implements IClientCore {
       tokenAddress,
       reference,
     ] = ClientCore.createDepositParameters(params);
+
+    // Depositing an ERC20 token?
+    if (tokenAddress !== AddressZero) {
+      const governanceERC20Instance = GovernanceERC20__factory.connect(
+        tokenAddress,
+        this.connectedSigner
+      );
+
+      const currentAllowance = await this.connectedSigner
+        .getAddress()
+        .then(address =>
+          governanceERC20Instance.allowance(address, daoAddress)
+        );
+      yield {
+        key: DaoDepositSteps.CHECKED_ALLOWANCE,
+        allowance: currentAllowance.toBigInt(),
+      };
+
+      if (currentAllowance.lt(amount)) {
+        const increaseAllowanceTx = await governanceERC20Instance.increaseAllowance(
+          daoAddress,
+          BigNumber.from(amount)
+        );
+        yield {
+          key: DaoDepositSteps.INCREASING_ALLOWANCE,
+          txHash: increaseAllowanceTx.hash,
+        };
+
+        await increaseAllowanceTx.wait().then(cr => {
+          if (
+            BigNumber.from(amount).gt(
+              cr.events?.find(e => e?.event === "Approval")?.args?.value
+            )
+          ) {
+            throw new Error("Could not increase allowance");
+          }
+        });
+
+        yield {
+          key: DaoDepositSteps.INCREASED_ALLOWANCE,
+          allowance: amount.toBigInt(),
+        };
+      }
+    }
+
+    // Doing the transfer
 
     const daoInstance = DAO__factory.connect(daoAddress, this.connectedSigner);
 
@@ -299,19 +349,91 @@ export abstract class ClientCore implements IClientCore {
       }
     }
 
-    return daoInstance
-      .deposit(tokenAddress, amount, reference, override)
-      .then((tx) => tx.wait())
-      .then((cr) => {
-        const eventAmount = cr.events?.find((e) => e?.event === "Deposited")
-          ?.args
-          ?.amount;
-        if (!amount.eq(eventAmount)) {
-          throw new Error(
-            `Deposited amount mismatch. Expected: ${amount.toBigInt()}, received: ${eventAmount.toBigInt()}`,
-          );
-        }
-      });
+    const depositTx = await daoInstance.deposit(
+      tokenAddress,
+      amount,
+      reference,
+      override
+    );
+    yield { key: DaoDepositSteps.DEPOSITING, txHash: depositTx.hash };
+
+    await depositTx.wait().then(cr => {
+      const eventAmount = cr.events?.find(e => e?.event === "Deposited")?.args
+        ?.amount;
+      if (!amount.eq(eventAmount)) {
+        throw new Error(
+          `Deposited amount mismatch. Expected: ${amount.toBigInt()}, received: ${eventAmount.toBigInt()}`
+        );
+      }
+    });
+    yield { key: DaoDepositSteps.DEPOSITED, amount: amount.toBigInt() };
+  }
+
+  protected estimateDeposit(params: IDeposit): Promise<IGasFeeEstimation> {
+    if (!this.connectedSigner)
+      throw new Error("A signer is needed for creating a DAO");
+
+    const [
+      daoAddress,
+      amount,
+      tokenAddress,
+      reference,
+    ] = ClientCore.createDepositParameters(params);
+
+    const daoInstance = DAO__factory.connect(daoAddress, this.connectedSigner);
+
+    const override =
+      tokenAddress !== AddressZero
+        ? {}
+        : {
+            value: amount,
+          };
+
+    const gasLimit = daoInstance.estimateGas.deposit(
+      tokenAddress,
+      amount,
+      reference,
+      override
+    );
+    return this.estimateGasFee(gasLimit);
+  }
+
+  protected estimateIncreaseAllowance(
+    tokenAddress: string,
+    daoAddress: string,
+    amount: bigint
+  ): Promise<IGasFeeEstimation> {
+    if (!this.connectedSigner)
+      throw new Error("A signer is needed for creating a DAO");
+
+    const governanceERC20Instance = GovernanceERC20__factory.connect(
+      tokenAddress,
+      this.connectedSigner
+    );
+
+    const gasLimit = governanceERC20Instance.estimateGas.increaseAllowance(
+      daoAddress,
+      BigNumber.from(amount)
+    );
+    return this.estimateGasFee(gasLimit);
+  }
+
+  protected currentAllowance(
+    tokenAddress: string,
+    daoAddress: string
+  ): Promise<bigint> {
+    if (!this.connectedSigner)
+      throw new Error("A signer is needed for creating a DAO");
+
+    const governanceERC20Instance = GovernanceERC20__factory.connect(
+      tokenAddress,
+      this.connectedSigner
+    );
+
+    return this.connectedSigner
+      .getAddress()
+      .then(address => governanceERC20Instance.allowance(address, daoAddress))
+      .then(allowance => allowance.toBigInt());
   }
 
   // IPFS METHODS